requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
pillow==9.5.0
numpy==1.24.3
opencv-python==4.8.0.74
matplotlib==3.7.2
scikit-learn==1.3.0
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
Flask==2.3.3
flask-cors==4.0.0
reportlab==4.0.7